+++ common-system-test/src/main/java/com/teamscale/test/commons/ExternalReport.java
@@ -1,26 +1,14 @@
 package com.teamscale.test.commons;
 
-/** Holds a single report that was uploaded to our fake Teamscale server. */
+/** Holds a single report uploaded to our fake Teamscale server. */
 public class ExternalReport {
 	private final String reportString;
-	private final String partition;
-	private final String format;
 
-	public ExternalReport(String reportString, String partition, String format) {
+	public ExternalReport(String reportString) {
 		this.reportString = reportString;
-		this.partition = partition;
-		this.format = format;
 	}
 
 	public String getReportString() {
 		return reportString;
-	}
-
-	public String getPartition() {
-		return partition;
-	}
-
-	public String getFormat() {
-		return format;
 	}
 }