+++ system-tests/tia-maven/maven-project/sub-project-B/sub-project-B1/src/test/java/org/example/IntegrationB1IT.java
@@ -1,0 +1,17 @@
+package org.example;
+
+import org.junit.jupiter.api.Test;
+import org.example.SUTB1;
+
+public class IntegrationB1IT {
+
+    @Test
+    public void itBlub() throws Exception {
+        new SUTB1().blub();
+    }
+
+    @Test
+    public void itGoo() {
+        new SUTB1().goo();
+    }
+}